pull request for merging dev to main

+++ index.js
@@ -16,7 +16,8 @@ const {
   signupValidation,
   isValidSession,
   isAdmin,
-  isLoggedIn
+  isLoggedIn,
+  reviewValidation
 } = require('./middleware');
 
 const saltRounds = 12;
@@ -108,7 +109,7 @@ app.get('/', async (req, res) => {
     req.session.uid = result[0]._id;
 
   }
-  res.render("index", { isLoggedIn: isLoggedIn(req) });
+  res.render("index", { isLoggedIn: isLoggedIn(req), username: req.session.username });
 });
 
 
@@ -162,6 +163,7 @@ app.get('/search-results', async (req, res) => {
       level,
       rating,
       sort,
+      username: req.session.username,
     });
 
   } catch (error) {
@@ -178,7 +180,7 @@ app.get('/course-details', async (req, res) => {
 
   const userBookmarks = await bookmarkCollection.find({ userId: userId }).toArray();
   const email = req.session.email;
-  
+
   const courseInfo = await courseCollection.findOne({ _id: new ObjectId(courseId) });
 
   let reviewSliderPairs = await Promise.all(
@@ -260,6 +262,7 @@ app.get('/course-details', async (req, res) => {
     CourslaRating: CourslaRating,
     userBookmarks,
     easterEgg: false,
+    myReviewPage: false
   });
 });
 
@@ -332,7 +335,7 @@ app.get('/bookmarks', async (req, res) => {
 
     const userBookmarks = await bookmarkCollection.find({ userId: userId }).toArray();
 
-    res.render('bookmarks', { bookmarkedCourses, isLoggedIn: isLoggedIn(req), userBookmarks });
+    res.render('bookmarks', { bookmarkedCourses, isLoggedIn: isLoggedIn(req), userBookmarks, username: req.session.username});
     //res.render('bookmarks', { isLoggedIn: isLoggedIn(req), userBookmarks });
     //console.log('userBookmarks', userBookmarks)
   } catch (error) {
@@ -569,16 +572,11 @@ app.get('/reviews/write/:courseid', async (req, res) => {
   const avatar = req.session.avatar;
   const courseId = req.params.courseid.replace(':', '');
 
-
-  // console.log(username);
   if (username == null) {
 
     req.session.courseId = courseId;
     return res.redirect("/login");
   }
-
-
-  // console.log("courseid here (2)", courseId)
 
   const reviews = await reviewCollection.find().toArray();
 
@@ -599,19 +597,13 @@ app.get('/reviews/write/:courseid', async (req, res) => {
     studentSupportSliderValue: review.StudentSupportRating
   }));
 
-  // const editReview = req.params.editReview === '/updateReview';
-
   // Find the specific review for the current user
   const specificReview = reviews.find(review => review.username === username && review.CourseID === courseId);
   const hasReview = Boolean(specificReview);
 
   if (specificReview) {
-    console.log("existing review", specificReview);
+
     const reviewId = specificReview._id.toString();
-
-    console.log("review id", reviewId)
-    console.log("do you?", hasReview);
-
     const renderData = {
       req: req,
       sliderValues: sliderValues,
@@ -622,7 +614,8 @@ app.get('/reviews/write/:courseid', async (req, res) => {
       specificReview: specificReview,
       hasReview: hasReview,
       reviewId: reviewId,
-      avatar: avatar
+      avatar: avatar,
+      // myReviewPage: myReviewPage
     }
     // console.log(avatar)
     res.render("write-review", renderData);
@@ -638,8 +631,8 @@ app.get('/reviews/write/:courseid', async (req, res) => {
       courseId: courseId,
       editReview: false,
       // specificReview: specificReview,
-      hasReview: hasReview
-      // reviewId: reviewId
+      hasReview: hasReview,
+
     };
     res.render("write-review", renderData);
   }
@@ -653,6 +646,9 @@ app.get('/reviews/write/updateReview/:id', async (req, res) => {
   const avatar = req.session.avatar;
   const reviewId = req.params.id; // Get the review ID from the URL parameter
   const reviews = await reviewCollection.find().toArray();
+  const myReviewPage = req.query.myReviewPage;
+
+  // console.log("hello?", myReviewPage)
 
   // console.log(username);
   const reviewSliderPairs = reviews.map(review => ({
@@ -683,7 +679,8 @@ app.get('/reviews/write/updateReview/:id', async (req, res) => {
     reviewId: reviewId,
     courseId: courseID,
     hasReview: false,
-    avatar: avatar
+    avatar: avatar,
+    myReviewPage: myReviewPage
   };
 
   res.render("write-review", renderData);
@@ -694,21 +691,27 @@ app.get('/reviews/write/updateReview/:id', async (req, res) => {
 //delete the review from database
 app.post('/reviews/deleteReview/:id', async (req, res) => {
   const courseId = req.params.id;
-
+  console.log("deleted review is for this course: ", courseId);
   // Get the review ID before deleting the review
   const deletedReview = await reviewCollection.findOne({ CourseID: courseId });
-  const deletedReviewId = deletedReview._id.toString();
 
-  // Delete the review from the database based on the review ID
-  await reviewCollection.deleteOne({ CourseID: courseId });
+  console.log(deletedReview);
 
-  // Delete the corresponding review ID from the array in the user document
-  await userCollection.updateOne(
-    { ReviewID: deletedReviewId },
-    { $pull: { ReviewID: deletedReviewId } }
-  );
+  if (deletedReview) {
+    const deletedReviewId = deletedReview._id.toString();
+    await reviewCollection.deleteOne({ CourseID: courseId });
 
-  res.redirect(`/course-details?courseId=${courseId}`);
+    // Delete the corresponding review ID from the array in the user document
+    await userCollection.updateOne(
+      { ReviewID: deletedReviewId },
+      { $pull: { ReviewID: deletedReviewId } }
+    );
+
+    res.redirect(`/course-details?courseId=${courseId}`);
+  }
+  else {
+    res.status(404).send('Review not found');
+  }
 });
 
 
@@ -720,6 +723,10 @@ app.post('/submitReview/:id', async (req, res) => {
   const courseId = req.params.id;
   const Existingreviews = await reviewCollection.findOne({ CourseID: courseId, email: email });
 
+  const myReviewPage = req.query.myReviewPage;
+
+  console.log('submit', myReviewPage)
+
   const { review,
     courseContentSliderValue,
     courseStructureSliderValue,
@@ -727,6 +734,8 @@ app.post('/submitReview/:id', async (req, res) => {
     studentSupportSliderValue,
     currentDate } = req.body;
 
+  // const errorMessage = await reviewValidation(req.body);
+  console.log(req.errorMessage);
 
   // if there is a exisiting review, direct user to edit their existing review
   if (Existingreviews) {
@@ -761,14 +770,11 @@ app.post('/submitReview/:id', async (req, res) => {
       CourseID: courseId
     });
 
-
+    console.log(review);
     const insertedReview = await reviewCollection.findOne({
       Review: review,
       email: email
     });
-    console.log(email);
-    // console.log("reviews", insertedReview);
-    console.log("inserted review", insertedReview._id.toString());
 
     const reviewCount = await reviewCollection.countDocuments({
       email: email
@@ -805,7 +811,7 @@ app.post('/submitReview/:id', async (req, res) => {
 
       const response = await userCollection.updateOne(
         { _id: new ObjectId(uid) },
-        {$set: {Badges: "Reviewer"}}
+        { $set: { Badges: "Reviewer" } }
       );
       // console.log('response',response)
 
@@ -815,19 +821,24 @@ app.post('/submitReview/:id', async (req, res) => {
       // to make sure they have badge
       await userCollection.updateOne(
         { _id: new ObjectId(uid) },
-        {$set: {Badges: "Reviewer"}}
+        { $set: { Badges: "Reviewer" } }
       )
     } else {
       // to prevent user remove review and get badge
       await userCollection.updateOne(
         { _id: new ObjectId(uid) },
-        {$set: {Badges: " "}}
+        { $set: { Badges: " " } }
       );
     }
 
   }
 
-  res.redirect('/course-details?courseId=' + courseId);
+  if (myReviewPage) {
+    res.redirect('/my-reviews');
+  }
+  else {
+    res.redirect('/course-details?courseId=' + courseId);
+  }
 });
 
 app.get('/profileReview', async (req, res) => {
@@ -856,33 +867,35 @@ app.get('/profileReview', async (req, res) => {
 });
 
 app.get('/my-reviews', async (req, res) => {
-  const courseId = req.query.courseId;
+  // const courseId = req.query.courseId;
   const reviews = await reviewCollection.find({ email: req.session.email }).toArray();
   // const reviews = await reviewCollection.find().toArray();
+  // console.log("email:", req.session.email);
   const username = req.session.username;
-  const userId = req.session.uid;
-  const email = req.session.email;
+  let courseId;
+  // const userId = req.session.uid;
+  // const email = req.session.email;
   const reviewGroups = {};
-
+  
   // Group reviews by courseId
   reviews.forEach(review => {
-    const courseId = review.CourseID;
+    courseId = review.CourseID;
     if (!reviewGroups[courseId]) {
       reviewGroups[courseId] = [];
     }
     reviewGroups[courseId].push(review);
+    // console.log("courseId: ", reviewGroups);
   });
-
+  
   const reviewSliderPairs = [];
 
-  const courseInfo = await courseCollection.findOne({ _id: new ObjectId(courseId) });
+  //const courseInfo = await courseCollection.findOne({ _id: new ObjectId(courseId) });
 
   // Retrieve course information for each group
   for (const courseId in reviewGroups) {
     const courseInfo = await courseCollection.findOne({ _id: new ObjectId(courseId) });
-
+    
     const groupReviews = reviewGroups[courseId];
-
     const groupSliderPairs = await Promise.all(
       groupReviews.map(async review => {
         const sliderValue = {
@@ -942,17 +955,18 @@ app.get('/my-reviews', async (req, res) => {
     );
   }
 
-  console.log(reviewSliderPairs)
+  // console.log(reviewSliderPairs)
   updateCourse(courseId, overallCategorySums, reviewSliderPairs.length);
 
   res.render("my-review", {
     req: req,
-    courseId: courseId,
+    // courseId: courseId,
     reviewSliderPairs: reviewSliderPairs,
     username: username,
     isLoggedIn: isLoggedIn(req),
     overallCategorySums: overallCategorySums,
-    Totalvote: totalvote
+    Totalvote: totalvote,
+    myReviewPage: true
   });
 });
